+++ src/constants.jl
@@ -1,14 +1,3 @@
-module SDF
-
-export header, Header,
-    file_summary,
-    Stagger, CellCentre, FaceX, FaceY, FaceZ, EdgeX, EdgeY, EdgeZ, Vertex,
-    AbstractBlockHeader,
-    PlainVariableBlockHeader,
-    PointVariableBlockHeader,
-    PlainMeshBlockHeader,
-    PointMeshBlockHeader
-
 const ID_LENGTH = 32
 const ENDIANNESS = 16911887
 const SDF_VERSION = 1
@@ -56,47 +45,3 @@ const DATATYPE_REAL16 = Int32(5)
 const DATATYPE_CHARACTER = Int32(6)
 const DATATYPE_LOGICAL = Int32(7)
 const DATATYPE_OTHER = Int32(8)
-
-include("sdf_header.jl")
-include("read_header.jl")
-include("read_data.jl")
-
-@generated function typemap(data_type::Val{N}) :: DataType where N
-    if N == DATATYPE_NULL
-        Nothing
-    elseif N == DATATYPE_INTEGER4
-        Int32
-    elseif N == DATATYPE_INTEGER8
-        Int64
-    elseif N == DATATYPE_REAL4
-        Float32
-    elseif N == DATATYPE_REAL8
-        Float64
-    elseif N == DATATYPE_REAL16
-        Nothing # unsupported
-    elseif N == DATATYPE_CHARACTER
-        Char
-    elseif N == DATATYPE_LOGICAL
-        Bool
-    else
-        Nothing
-    end
-end
-
-file_summary(filename) = open(file_summary, filename)[2]
-
-function file_summary(f::IOStream)
-    h = header(f)
-    blocks = Dict{Symbol,AbstractBlockHeader}()
-
-    block_start = h.summary_location
-    for i in Base.OneTo(h.nblocks)
-        block = BlockHeader(f, block_start, h.string_length, h.block_header_length)
-        blocks = push!(blocks, Symbol(block.id) => read(f, block))
-        block_start = block.next_block_location
-    end
-
-    h, blocks
-end
-
-end  # module SDF
