+++ src/utils.jl
@@ -4,8 +4,3 @@ end
 
 get_normalization(::Data, block) = block.mult
 get_normalization(::Grid, block) = block.mults
-
-# https://github.com/JuliaLang/Statistics.jl/pull/28
-@static if VERSION < v"1.6.0"
-    Statistics.middle(x::Number, y::Number) = x/2 + y/2
-end
